// src/types/IMovie.ts
export interface Movie {
  id: number;
  title: string;
  originalTitle: string;
  language:	string;
  relaseYear:	number;
  releaseDate: string;
  genres:	string[];
  plot:	string;
  runtime: number;
  budget:	string;
  revenue	:string;
  homepage:	string;
  status:	string;
  posterUrl: string;
  backdropUrl: string;
  trailerUrl:	string;
  trailerYouTubeId:	string;
  tmdbRating:	number;
  searchL: string;
  keywords: string[];
  countriesOfOrigin: string[];
  languages: string[];
  cast: string[];
  director: string;
  production: string;
  awardsSummary: string;
}